# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "字节青训营-模仿抖音项目",
        "title": "mock tiktok",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "联系人",
            "url": "http://www.swagger.io/support",
            "email": "584807419@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0 版本"
    },
    "basePath": "/",
    "paths": {
        "/douyin/relation/action": {
            "post": {
                "description": "关注或者取消关注用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Relation"
                ],
                "summary": "关注操作",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "关注者",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "被关注者",
                        "name": "to_user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "参数不正确",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "403": {
                        "description": "未登录",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/douyin/relation/follow/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Relation"
                ],
                "summary": "关注列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/douyin/relation/follower/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Relation"
                ],
                "summary": "粉丝列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/douyin/relation/friend/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "好友列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}